Updates for QICoreMeasureExport.cy.ts HR verification
