+++ cypress/Shared/MeasureCQL.ts
@@ -37,6 +37,64 @@ export class MeasureCQL {
         'where ValidEncounter.period during "Measurement Period"\n' +
         'define "Initial PopulationOne":\n' +
         'true\n'
+
+    public static readonly zipfileExportQICore = 'library TestLibrary1678378360032 version \'0.0.000\'\n' +
+        '\n' +
+        'using QICore version \'4.1.1\'\n' +
+        '\n' +
+        'include FHIRHelpers version \'4.1.000\' called FHIRHelpers\n' +
+        'include SupplementalDataElements version \'3.1.000\' called SDE\n' +
+        'include CQMCommon version \'1.0.000\' called CQMCommon \n' +
+        'include FHIRCommon version \'4.1.000\' called FHIRCommon\n' +
+        '\n' + '\n' +
+        'codesystem "SNOMEDCT": \'http://snomed.info/sct\' \n' +
+        'codesystem "ActCode": \'http://terminology.hl7.org/CodeSystem/v3-ActCode\'  \n' +
+        '\n' + '\n' + '\n' +
+        'valueset "Care Services in Long-Term Residential Facility": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1014\' \n' +
+        'valueset "Diabetic Retinopathy": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.327\' \n' +
+        'valueset "Level of Severity of Retinopathy Findings": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1283\' \n' +
+        'valueset "Macular Edema Findings Present": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1320\' \n' +
+        'valueset "Macular Exam": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1251\' \n' +
+        'valueset "Medical Reason": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1007\' \n' +
+        'valueset "Nursing Facility Visit": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1012\' \n' +
+        'valueset "Office Visit": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1001\' \n' +
+        'valueset "Ophthalmological Services": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1285\' \n' +
+        'valueset "Patient Reason": \'http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1008\' \n' +
+        '\n' +
+        'code "Healthcare professional (occupation)": \'223366009\' from "SNOMEDCT" display \'Healthcare professional (occupation)\'\n' +
+        'code "Medical practitioner (occupation)": \'158965000\' from "SNOMEDCT" display \'Medical practitioner (occupation)\'\n' +
+        'code "Ophthalmologist (occupation)": \'422234006\' from "SNOMEDCT" display \'Ophthalmologist (occupation)\'\n' +
+        'code "Optometrist (occupation)": \'28229004\' from "SNOMEDCT" display \'Optometrist (occupation)\'\n' +
+        'code "Physician (occupation)": \'309343006\' from "SNOMEDCT" display \'Physician (occupation)\'\n' +
+        'code "virtual": \'VR\' from "ActCode" display \'virtual\'\n' +
+        'code "Macular edema absent (situation)": \'428341000124108\' from "SNOMEDCT" display \'Macular edema absent (situation)\'\n' +
+        'code "AMB" : \'AMB\' from "ActCode" display \'Ambulatory\'\n' +
+        '\n' +
+        'parameter "Measurement Period" Interval<DateTime>\n' +
+        '\n' +
+        'context Patient\n' +
+        '\n' +
+        'define "SDE Ethnicity":\n' +
+        '  SDE."SDE Ethnicity"\n' +
+        '\n' +
+        'define "SDE Payer":\n' +
+        '  SDE."SDE Payer"\n' +
+        '\n' +
+        'define "SDE Race":\n' +
+        '  SDE."SDE Race"\n' +
+        '\n' +
+        'define "SDE Sex":\n' +
+        '  SDE."SDE Sex"\n' +
+        '\n' + '\n' + '\n' +
+        'define "ipp":\n' +
+        'true\n' +
+        '\n' +
+        'define "num":\n' +
+        'true\n' +
+        '\n' +
+        'define "denom":\n' +
+        'true'
+
     public static readonly ICFCleanTestQICore_CQL_without_using = 'library SimpleFhirLibrary version \'0.0.004\'\n' +
         'include FHIRHelpers version \'4.1.000\' called FHIRHelpers\n' +
         'codesystem \"SNOMEDCT:2017-09\": \'http://snomed.info/sct/731000124108\' version \'http://snomed.info/sct/731000124108/version/201709\'\n' +
