--- cypress/fixtures/QiCoreHRCompare/GroupScoringAndPCSections.html
@@ -1,54 +1,0 @@
-<th scope="row" class="row-header">Scoring</th>
-
-<td class="content-container">[http://terminology.hl7.org/CodeSystem/measure-scoring#proportion: 'Proportion']</td>
-</tr>
-
-
-
-  
-<tr>
-
-<th scope="row" class="row-header">Scoring Unit</th>
-
-<td class="content-container">[https://clinicaltables.nlm.nih.gov/#ml: 'ml milliLiters']</td>
-</tr>
-
-
-
-
-
-
-  
-<tr>
-
-<th scope="row" class="row-header">Type</th>
-
-<td class="content-container">[http://terminology.hl7.org/CodeSystem/measure-type#outcome: 'Outcome']</td>
-</tr>
-
-
-
-
-
-  
-<tr>
-
-<th scope="row" class="row-header">Improvement Notation</th>
-
-<td class="content-container">[http://terminology.hl7.org/CodeSystem/measure-improvement-notation#increase: 'Increased score indicates improvement']</td>
-</tr>
-
-
-
-
-
-
-<tr>
-
-<th colspan="2" scope="row" class="row-header">Measure Logic</th>
-
-</tr>
-
-<tr>
-
-<th scope="row" class="row-header">Primary Library</th>
