--- cypress/fixtures/QiCoreHRCompare/LiquidStyleSection.html
@@ -1,83 +1,0 @@
-        /*
-  ===============================
-  Liquid Narrative Styling Start
-  ===============================
-  */
-
-        .narrative-table {
-            border: 1pt inset rgb(0, 0, 0);
-            line-height: 10pt;
-            border-collapse: separate;
-            -webkit-border-horizontal-spacing: 2px;
-            -webkit-border-vertical-spacing: 2px;
-        }
-
-        .narrative-table>tbody {
-            display: table-row-group;
-            vertical-align: middle;
-        }
-
-        .narrative-table>tbody>tr {
-            display: table-row;
-            background-color: lightgray;
-            vertical-align: inherit;
-        }
-
-        .narrative-table>tbody>tr>th {
-            padding: 5px;
-        }
-
-        .section-header {
-            background-color: #656565;
-            color: whitesmoke;
-            vertical-align: inherit;
-        }
-
-        .row-header {
-            background-color: #656565;
-            color: whitesmoke;
-            vertical-align: inherit;
-        }
-
-        .tab-one {
-            padding-left: 5px;
-        }
-
-        @media (min-width: 768px) {
-            .row-header {
-                min-width: 138px;
-                max-width: 138px;
-            }
-
-            .content-container {
-                max-width: 576px;
-            }
-        }
-
-        @media (min-width: 992px) {
-            .row-header {
-                min-width: 181px;
-                max-width: 181px;
-            }
-
-            .content-container {
-                max-width: 745px;
-            }
-        }
-
-        @media (min-width: 1100px) {
-            .row-header {
-                min-width: 213px;
-                max-width: 213px;
-            }
-
-            .content-container {
-                max-width: 873px;
-            }
-        }
-
-        /*
-        ===============================
-        Liquid Narrative Styling End
-        ===============================
-        */
