--- cypress/fixtures/QiCoreHRCompare/QualifyingEncountersSection.html
@@ -1,28 +1,0 @@
-
-</tr>
-<tr>
-
-  <td>
-    <pre style="border: none;" class="content-container highlight language-cql"><code class="language-cql">define &quot;Qualifying Encounters&quot;:
-  ( [Encounter: &quot;Office Visit&quot;]
-    union [Encounter: &quot;Annual Wellness Visit&quot;]
-    union [Encounter: &quot;Preventive Care Services - Established Office Visit, 18 and Up&quot;]
-    union [Encounter: &quot;Preventive Care Services-Initial Office Visit, 18 and Up&quot;]
-    union [Encounter: &quot;Home Healthcare Services&quot;] ) ValidEncounter
-    where ValidEncounter.period during &quot;Measurement Period&quot;</code></pre>
-  </td>
-
-</tr>
-        
-<tr>
-  <th scope="row" rowspan="2" class="row-header">
-    
-      
-      <a name="fhirhelpers-tointerval"> </a>
-    
-    Logic Definition
-  </th>
-
-  <td class="content-container"><em>Library Name:</em> FHIRHelpers</td>
-
-</tr>
