--- cypress/fixtures/QiCoreHRCompare/TerminologyAndOtherDependenciesSections.html
@@ -1,401 +1,0 @@
-</td>
-</tr>
-
-
-
-
-  
-  
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Library FHIRHelpers
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: https://madie.cms.gov/Library/FHIRHelpers|4.1.000
-    <br/>
-    <em>Canonical URL</em>: <tt>https://madie.cms.gov/Library/FHIRHelpers|4.1.000</tt>
-    
-  </td>
-</tr>
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Value set Office Visit
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1001
-    <br/>
-    <em>Canonical URL</em>: <tt>http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1001</tt>
-    
-  </td>
-</tr>
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Value set Annual Wellness Visit
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1240
-    <br/>
-    <em>Canonical URL</em>: <tt>http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1240</tt>
-    
-  </td>
-</tr>
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Value set Preventive Care Services - Established Office Visit, 18 and Up
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1025
-    <br/>
-    <em>Canonical URL</em>: <tt>http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1025</tt>
-    
-  </td>
-</tr>
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Value set Preventive Care Services-Initial Office Visit, 18 and Up
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1023
-    <br/>
-    <em>Canonical URL</em>: <tt>http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1023</tt>
-    
-  </td>
-</tr>
-
-<tr>
-  
-  
-  
-  
-  
-<th scope="row" class="row-header">Dependency</th>
-
-  
-  <td class="content-container">
-    
-    <em>Description</em>: Value set Home Healthcare Services
-    
-    <br/>
-    
-    
-    
-    
-    
-    
-    <em>Resource</em>: http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1016
-    <br/>
-    <em>Canonical URL</em>: <tt>http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1016</tt>
-    
-  </td>
-</tr>
-
-
-  
-  
-  <tr>
-  
-<th scope="row" class="row-header">Parameter</th>
-
-  <td class="content-container">
-    <em>Name</em>: Measurement Period
-    <br/>
-    <em>Use</em>: in
-    <br/>
-    <em>Min Cardinality</em>: 0
-    <br/>
-    <em>Max Cardinality</em>: 1
-    <br/>
-    <em>Type</em>: Period
-  </td>
-</tr>
-  
-  <tr>
-  
-<th scope="row" class="row-header">Parameter</th>
-
-  <td class="content-container">
-    <em>Name</em>: Qualifying Encounters
-    <br/>
-    <em>Use</em>: out
-    <br/>
-    <em>Min Cardinality</em>: 0
-    <br/>
-    <em>Max Cardinality</em>: *
-    <br/>
-    <em>Type</em>: Resource
-  </td>
-</tr>
-  
-  
-  <tr>
-    <th colspan="2" scope="row" class="section-header"><a name="effective-data-requirements"> </a>Measure Logic Data Requirements</th>
-  </tr>
-  
-  
-<tr>
-  <th scope="row" class="row-header">Data Requirement</th>
-  <td class="content-container">
-    <em>Type</em>: Encounter
-    <br/>
-  
-    <em>Profile(s)</em>: 
-  
-    http://hl7.org/fhir/us/qicore/StructureDefinition/qicore-encounter
-    <br/>        
-  
-   
-   
-    <em>Must Support Elements</em>: type, period
-    <br/>
-   
-  
-    <em>Code Filter(s)</em>: 
-    <br/>
-  
-  
-    <span class="tab-one"><em>Path</em>: type</span>
-    <br/>
-  
-  
-  
-    <span class="tab-one"><em>ValueSet</em>:</span> http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1001
-    <br/> 
-  
-  
-  
-   
-  </td>
-</tr>
-
-<tr>
-  <th scope="row" class="row-header">Data Requirement</th>
-  <td class="content-container">
-    <em>Type</em>: Encounter
-    <br/>
-  
-    <em>Profile(s)</em>: 
-  
-    http://hl7.org/fhir/us/qicore/StructureDefinition/qicore-encounter
-    <br/>        
-  
-   
-   
-    <em>Must Support Elements</em>: type, period
-    <br/>
-   
-  
-    <em>Code Filter(s)</em>: 
-    <br/>
-  
-  
-    <span class="tab-one"><em>Path</em>: type</span>
-    <br/>
-  
-  
-  
-    <span class="tab-one"><em>ValueSet</em>:</span> http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.526.3.1240
-    <br/> 
-  
-  
-  
-   
-  </td>
-</tr>
-
-<tr>
-  <th scope="row" class="row-header">Data Requirement</th>
-  <td class="content-container">
-    <em>Type</em>: Encounter
-    <br/>
-  
-    <em>Profile(s)</em>: 
-  
-    http://hl7.org/fhir/us/qicore/StructureDefinition/qicore-encounter
-    <br/>        
-  
-   
-   
-    <em>Must Support Elements</em>: type, period
-    <br/>
-   
-  
-    <em>Code Filter(s)</em>: 
-    <br/>
-  
-  
-    <span class="tab-one"><em>Path</em>: type</span>
-    <br/>
-  
-  
-  
-    <span class="tab-one"><em>ValueSet</em>:</span> http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1025
-    <br/> 
-  
-  
-  
-   
-  </td>
-</tr>
-
-<tr>
-  <th scope="row" class="row-header">Data Requirement</th>
-  <td class="content-container">
-    <em>Type</em>: Encounter
-    <br/>
-  
-    <em>Profile(s)</em>: 
-  
-    http://hl7.org/fhir/us/qicore/StructureDefinition/qicore-encounter
-    <br/>        
-  
-   
-   
-    <em>Must Support Elements</em>: type, period
-    <br/>
-   
-  
-    <em>Code Filter(s)</em>: 
-    <br/>
-  
-  
-    <span class="tab-one"><em>Path</em>: type</span>
-    <br/>
-  
-  
-  
-    <span class="tab-one"><em>ValueSet</em>:</span> http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1023
-    <br/> 
-  
-  
-  
-   
-  </td>
-</tr>
-
-<tr>
-  <th scope="row" class="row-header">Data Requirement</th>
-  <td class="content-container">
-    <em>Type</em>: Encounter
-    <br/>
-  
-    <em>Profile(s)</em>: 
-  
-    http://hl7.org/fhir/us/qicore/StructureDefinition/qicore-encounter
-    <br/>        
-  
-   
-   
-    <em>Must Support Elements</em>: type, period
-    <br/>
-   
-  
-    <em>Code Filter(s)</em>: 
-    <br/>
-  
-  
-    <span class="tab-one"><em>Path</em>: type</span>
-    <br/>
-  
-  
-  
-    <span class="tab-one"><em>ValueSet</em>:</span> http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.464.1003.101.12.1016
-    <br/> 
-  
-  
-  
-   
-  </td>
-</tr>
-
-  
-  <tr>
-
-<th colspan="2" scope="row" class="row-header">Measure Logic Definitions</th>
-
-</tr>
-  
-          
-<tr>
-  <th scope="row" rowspan="2" class="row-header">
-    
-
