--- cypress/fixtures/QiCoreHRCompare/VersionSection.html
@@ -1,10 +1,0 @@
-
-
-
-
-<tr>
-
-<th scope="row" class="row-header">Version Number</th>
-
-<td class="content-container">Draft based on 0.0.000</td>
-</tr>
