--- cypress/fixtures/QiCoreHRCompare/VersionSection_ver1.html
@@ -1,6 +1,0 @@
-<tr>
-
-<th scope="row" class="row-header">Version Number</th>
-
-<td class="content-container">1.0.000</td>
-</tr>
